feat: use BlogLayout as the default layout of all .svx files
