+++ src/routes/generate-and-put-in-context-or-compute-from-context-externally-xstate.svx
@@ -1,0 +1,16 @@
+---
+title: Generate and put in context or compute from context externally XState
+description: Lorem ipsum dolor sit amet, consectetur adipisicing elit. In delectus, aliquid beatae officia placeat iusto sint cumque corporis deserunt magnam ad voluptatibus, culpa obcaecati accusamus, quidem recusandae. Qui, nulla error!
+datetime: 23/01/2000
+tags: [Front-end, Temporal]
+
+customText: eheh
+---
+
+<script>
+    const lol = 'lol'
+</script>
+
+Hi **eheh**
+
+{customText} {lol}
