--- src/routes/test.svx
@@ -1,7 +1,0 @@
-<script lang="ts">
-    const customText: string = 'eheh'
-</script>
-
-Hi **eheh**
-
-{customText}
